requirements.txt
flake8==3.5.0
mccabe==0.6.1
numpy==1.14.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
